+++ CHANGELOG.md
@@ -9,6 +9,12 @@
 * Add benchmark to run DCP Loading Workloads (#357)
 * Add thread_count parameter to S3StorageWriter (#370)
 
+## New features
+* Shadow Copies Implemented
+
+## New features
+* Shadow Copies Implemented
+
 ## v1.4.3 (July 25, 2025)
 
 ### New features